Fold the MAC instruction into ai_ext to keep the custom ISA extension implementation in a single location.
