+++ src/arch/riscv/insts/SConscript
@@ -40,4 +40,3 @@ Source('vector.cc', tags=['riscv isa'])
 Source('zcmp.cc', tags=['riscv isa'])
 Source('zcmt.cc', tags=['riscv isa'])
 Source('ai_ext.cc', tags=['riscv isa'])
-Source('mac.cc', tags=['riscv isa'])
