+++ src/arch/riscv/isa/decoder.isa
@@ -2612,7 +2612,7 @@ decode QUADRANT default Unknown::unknown() {
                 }
             }
         }
-        
+
         // Using custom-0 opcode space (0x02)
         // Encoding group (custom-0, opcode=0x0B):
         // - R-type (funct3=0x0):
