+++ src/arch/riscv/isa/includes.isa
@@ -50,7 +50,6 @@ output header {{
 #include "arch/riscv/insts/bs.hh"
 #include "arch/riscv/insts/compressed.hh"
 #include "arch/riscv/insts/ai_ext.hh"
-#include "arch/riscv/insts/mac.hh"
 #include "arch/riscv/insts/mem.hh"
 #include "arch/riscv/insts/pseudo.hh"
 #include "arch/riscv/insts/standard.hh"
